.claude/skills/obsidian-rag/scripts/requirements.txt
chromadb>=0.4.0
pyyaml>=6.0
pytest>=7.0.0
